servers/requirements/production.txt
GitPython==0.3.2.RC1
PyYAML==3.10
async==0.6.1
gitdb==0.5.4
smmap==0.8.2
tornado==2.2
